% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{airpred.predict}
\alias{airpred.predict}
\title{Generate predictions from a previously trained model}
\usage{
airpred.predict(prepped = T)
}
\description{
Generate predictions from a previously trained model
}
