% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{airpred_clean}
\alias{airpred_clean}
\title{10 days of PM2.5 prediction input data}
\format{21560 observations and 116 variables.}
\usage{
airpred_clean
}
\description{
A dataset containing the data for all PM2.5 monitors in the US from Jan. 1st 2012 to
Jan. 10th 2012. These data have already been transformed, normalized, and imputed.
}
\keyword{datasets}
