% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{airpred_monitor_locations}
\alias{airpred_monitor_locations}
\title{WGS84 Lat and Lon locations of the PM2.5 monitors}
\format{2156 observations and 3 variables.}
\usage{
airpred_monitor_locations
}
\description{
Locations of all the PM2.5 monitors, can be linked using the site number variable
}
\keyword{datasets}
