% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_params.R
\name{clean_model_configs}
\alias{clean_model_configs}
\title{Remove model config file from current directory}
\usage{
clean_model_configs()
}
\description{
Remove model config file from current directory
}
