% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{destandardize_all}
\alias{destandardize_all}
\title{Destandardize Data}
\usage{
destandardize_all(info, store = FALSE, load = TRUE)
}
\arguments{
\item{info}{the data set to be normalized}

\item{store}{should the normalization values be stored}

\item{load}{should the normalization values be loaded}
}
\value{
data frame
}
\description{
convert variables from z-scores to original values
}
