% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{display_config}
\alias{display_config}
\title{Print a config file's contents to the console}
\usage{
display_config(path = "config.yml")
}
\arguments{
\item{path}{the path to the config file you want to print.}
}
\description{
Print a config file's contents to the console
}
