% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_params.R
\name{edit_params}
\alias{edit_params}
\title{Generate config files for all training models}
\usage{
edit_params(path = ".")
}
\arguments{
\item{path}{the directory that will store the generated config files}
}
\description{
Generate config files for all training models
}
\seealso{
gen_model_config
}
