% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_params.R
\name{ensemble_config_check}
\alias{ensemble_config_check}
\title{Check that the model parameters allow for the h2o ensemble to run}
\usage{
ensemble_config_check()
}
\description{
Check that the model parameters allow for the h2o ensemble to run
}
