% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train.R
\name{ensemble_formula}
\alias{ensemble_formula}
\title{Build formula for ensmeble model}
\usage{
ensemble_formula(models)
}
\arguments{
\item{models}{a list of trained models}
}
\value{
A string with the formula used in the ensemble model
}
\description{
returns a string with the form "MonitorData ~ s(model1) + s(model2) ...
}
