% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_config}
\alias{gen_config}
\title{Generate Config File Skeleton}
\usage{
gen_config(path = ".")
}
\arguments{
\item{path}{the directory that the config file should be saved in}
}
\value{
Null, but saves a yml file with the headers needed to run the prediction model saved
}
\description{
Generate Config File Skeleton
}
\details{
\itemize{
  \item{\code{Monitor}} {The pollution type the data will be trained on}
  \item{\code{Data_Location}} {The directory holding the required data files}
 }
}
