% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{gen_data_paths}
\alias{gen_data_paths}
\title{Generate Data File path}
\usage{
gen_data_paths(path = "../predictions/EPANO2")
}
\arguments{
\item{path}{the directory storing the data to be used in the model}
}
\value{
A list where each object is a variable containing a vector of all existing files
in the checked directory that match the file name pattern for that variable specified in
the Data_Location.yml file.
}
\description{
Generate Data File path
}
