% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{gen_grid_config}
\alias{gen_grid_config}
\title{Generate configuration file for use with the grid search function}
\usage{
gen_grid_config(default = T, path = ".")
}
\description{
Generate configuration file for use with the grid search function
}
