% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{gen_norm_vals}
\alias{gen_norm_vals}
\title{Generate Data Frame with stored metadata}
\usage{
gen_norm_vals(info, store = TRUE, load = FALSE)
}
\arguments{
\item{info}{A dataframe}

\item{store}{should the genreated metadata be stored}

\item{load}{SHould the metadata be loaded instead}
}
\value{
metadata dataframe
}
\description{
Generate Data Frame with stored metadata
}
