% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenWeights.R
\name{gen_QD_weights}
\alias{gen_QD_weights}
\title{Generate geographic weighting matrix}
\usage{
gen_QD_weights(orig, query, k = 4, threshold)
}
\arguments{
\item{orig}{the data}

\item{threshold}{}
}
\description{
Given two sets of points, generate a matrix of weights for the k nearest neighbors
for use in distance weighted averaging of values.
}
