% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{gen_stats}
\alias{gen_stats}
\title{Generate Data Frame with stored mean and sd}
\usage{
gen_stats(info, store = TRUE, load = FALSE)
}
\arguments{
\item{info}{A dataframe}

\item{store}{should the genreated metadata be stored}

\item{load}{SHould the metadata be loaded instead}
}
\value{
metadata dataframe. The first row is the mean of a
variable and the second is the standard deviation
}
\description{
Generate Data Frame with stored mean and sd
}
