% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenWeights.R
\name{gen_weights}
\alias{gen_weights}
\title{Generate geographic weighting matrix}
\usage{
gen_weights(orig, query, k = 4, threshold)
}
\arguments{
\item{orig}{the data in which the neighbors are being found}

\item{query}{the data for which the neighbors need to be found}

\item{k}{the number of neighbors being searched for}

\item{threshold}{}
}
\description{
Given two sets of points, generate a matrix of weights for the k nearest neighbors
for use in distance weighted averaging of values.
}
