% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interfaces.R
\name{get_csv_data}
\alias{get_csv_data}
\title{Generate training data from CSV}
\usage{
get_csv_data()
}
\value{
data frame
}
\description{
Given a csv file in the config file, generates a dataframe from the file ready to use for training
}
