% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interfaces.R
\name{get_rds_data}
\alias{get_rds_data}
\title{Generate training data from RDS}
\usage{
get_rds_data()
}
\value{
data frame
}
\description{
Given a rds file in the config file, generates a dataframe from the file ready to use for training
}
