% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{grid_to_config}
\alias{grid_to_config}
\title{Generate Model configs from grid search results}
\usage{
grid_to_config(grid_out, path = ".")
}
\description{
Generate Model configs from grid search results
}
