% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{h2o_impute_all}
\alias{h2o_impute_all}
\title{Impute all specified variables}
\usage{
h2o_impute_all(info, init = T, shutdown = T)
}
\arguments{
\item{info}{dataframe containing all variables for training}

\item{init}{should the h2o instance be initialized?
Only set as F if the h2o instance has already been initialized.}

\item{shutdown}{Should the h2o instance be shutdown after imputation, defaults to T.}
}
\value{
a data frame
}
\description{
Impute all specified variables
}
\examples{
\dontrun{
h2o.init()
info <- h2o_impute_all(info, init = F)
}
}
