% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{implemented_models}
\alias{implemented_models}
\title{List Implemented training models}
\usage{
implemented_models()
}
\value{
vector of implemented models
}
\description{
List Implemented training models
}
\details{
This function returns the acceptable strings to be included in the Training_Models
field in the config file. A string included in that section not currently in this list
will result in an error.

The currently implemented models are as follows
\itemize{
  \item{\code{nn}} {A neural network algorithm}
  \item{\code{forest}} {A random forest algorithm}
  \item{\code{gradboost}} {A gradient boost algorithm}
}
}
