% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{impute_all}
\alias{impute_all}
\title{Impute full dataset}
\usage{
impute_all(info)
}
\arguments{
\item{info}{the dataset being imputed}
}
\value{
data.table with replaced missing values
}
\description{
Impute full dataset
}
\details{
As a side effect of this function, all of the models used to impute the values
are saved in the directory specified in the configuration file. Ensure that there
is enough space available in that folder to store the generated models. These models
will later be used when data is being prepared for prediction.
}
