% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{impute_all_parallel}
\alias{impute_all_parallel}
\title{Parallel implementation of Imputation algorithm}
\usage{
impute_all_parallel(info)
}
\arguments{
\item{info}{the dataset being imputed}
}
\description{
Parallel implementation of Imputation algorithm
}
\seealso{
impute_all
}
