% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{join_data}
\alias{join_data}
\title{Join processed variables into a single data.table}
\usage{
join_data(files = NULL)
}
\arguments{
\item{files}{used when being called from process data, should not be used when called
from the console. Passes in the list of variables}
}
\value{
none, but saves the processed data as an RDS file
}
\description{
Join processed variables into a single data.table
}
