% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_data}
\alias{load_data}
\title{Load csv or RDS data}
\usage{
load_data(path)
}
\arguments{
\item{path}{the path of the data beint loaded

Load a data.frame like object stored as a csv or rds file. Passing any other
data type will return an error.}
}
\description{
Load csv or RDS data
}
