% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{load_predict_data}
\alias{load_predict_data}
\title{Load and prepare data for prediction}
\usage{
load_predict_data()
}
\description{
Uses \code{predict_mid_process} and \code{predict_data} from the
configuration file.
}
