% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_structure.R
\name{make_default_structure}
\alias{make_default_structure}
\title{Create airpred project directory structure}
\usage{
make_default_structure(path = ".")
}
\arguments{
\item{path}{directory serving as the airpred project. By default, the current working directory.}
}
\description{
Create airpred project directory structure
}
