% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{MLE_impute}
\alias{MLE_impute}
\title{Impute varibles using mixed linear models}
\usage{
MLE_impute(info, var)
}
\arguments{
\item{info}{data set}

\item{var}{the variable to be imputed}
}
\value{
a vector of the imputed values
}
\description{
Impute varibles using mixed linear models
}
