% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{normalize_all}
\alias{normalize_all}
\title{Normalize Data}
\usage{
normalize_all(info, store = TRUE, load = FALSE)
}
\arguments{
\item{info}{the data set to be normalized}

\item{store}{should the normalization values be stored}

\item{load}{should the normalization values be loaded}
}
\value{
data frame
}
\description{
Normalize Data
}
