% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_params.R
\name{param_config_check}
\alias{param_config_check}
\title{Check to see if model config files are in use}
\usage{
param_config_check(path = ".")
}
\value{
boolean
}
\description{
Check to see if model config files are in use
}
