% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{predict_impute_all}
\alias{predict_impute_all}
\title{Impute full dataset}
\usage{
predict_impute_all(info)
}
\arguments{
\item{info}{the dataset being imputed}
}
\value{
data.table with replaced missing values
}
\description{
Impute full dataset
}
\details{
This function relies on all variables having previously been imputed. If these models do not exist
or the directory in the configuration file is misspecified an error reporting that the file cannor be found
will be generated.
}
