% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{process_annual}
\alias{process_annual}
\title{process annual data}
\usage{
process_annual(files)
}
\arguments{
\item{files}{vector of file paths}
}
\value{
a Data Frame containing the value, dates, and site IDs for the given variable
}
\description{
process annual data
}
