% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{process_daily}
\alias{process_daily}
\title{Process Daily Data}
\usage{
process_daily(files)
}
\arguments{
\item{files}{vector of paths to files}
}
\value{
data frame
}
\description{
Process Daily Data
}
