% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{process_data}
\alias{process_data}
\title{Process all data for a given data set}
\usage{
process_data()
}
\arguments{
\item{path}{folder where the prediction data resides}
}
\value{
none
}
\description{
Process all data for a given data set
}
\details{
saves RDS files into the folder specified in the config file
}
