% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_all}
\alias{standardize_all}
\title{Standardize Data}
\usage{
standardize_all(info, store = TRUE, load = FALSE)
}
\arguments{
\item{info}{the data set to be normalized}

\item{store}{should the normalization values be stored}

\item{load}{should the normalization values be loaded}
}
\value{
data frame
}
\description{
standardize all variables to be z-scores
}
