% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train.R
\name{train}
\alias{train}
\title{Train Air Pollution Model}
\usage{
train(init = T, shutdown = T)
}
\arguments{
\item{init}{Boolean for whether or not an h2o cluster should be initiated on call of
this function.}

\item{shutdown}{Boolean for whether or not the h2o cluster should shutdown on termination}
}
\value{
NULL, but saves the models required to predict.
}
\description{
Train Air Pollution Model
}
