% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train.R
\name{train_ensemble}
\alias{train_ensemble}
\title{Train h2o ensemble model}
\usage{
train_ensemble(info, models, model_name)
}
\arguments{
\item{info}{h2o data training frame}

\item{models}{list of h2o trained models}

\item{model_name}{name with which to save the model}
}
\value{

}
\description{
Train h2o ensemble model
}
