% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train.R
\name{train_forest}
\alias{train_forest}
\title{Train Random Forest}
\usage{
train_forest(info, model_name)
}
\arguments{
\item{info}{data frame with model data}

\item{model_name}{name with which to save the model}
}
\value{
h2o model
}
\description{
Train Random Forest
}
