% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train.R
\name{train_generic}
\alias{train_generic}
\title{Train an h2o model using the generic architecture}
\usage{
train_generic(model, info)
}
\arguments{
\item{model}{the name of the function to run}

\item{info}{the data for use with the model}
}
\description{
Not currently developed
}
