% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform.R
\name{transform_all}
\alias{transform_all}
\title{Transform Data}
\usage{
transform_all(info, store = TRUE, load = FALSE)
}
\arguments{
\item{info}{the data set to be transformed}

\item{store}{should the transformation values be stored}

\item{load}{should the transformation values be loaded}
}
\value{
data frame
}
\description{
Transform Data
}
