% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{gen_config}
\alias{gen_config}
\title{Generate Config File Skeleton}
\usage{
gen_config(default = TRUE, path = ".", in_list = NULL)
}
\arguments{
\item{default}{A boolean determining whether or not
default values should be loaded into the generated config file}

\item{path}{the directory that the config file should be saved in}

\item{in_list}{An optional list that can be passed in to generate values for a config file.
Currently, the list is not validated, so the config file is not guaranteed to work with
the structure.}
}
\value{
Null, but saves a yml file with the headers needed to run the prediction model saved
}
\description{
Generate Config File Skeleton
}
\details{
The following are the items contained in the config file. All of them must be present in order for
the model to run successfully.

\itemize{
  \item{\code{impute}} {A boolean. If TRUE, airpred will generate imputations for
                        specified variables}
  \item{\code{impute_vars}} {Either "default" or a path to a yaml file listing the variables to be
                             imputed. If a custom file is being used, the top level of the
                             yaml file should be "base" with the elements of "base" being the
                             names of the variables being imputed.}
  \item{\code{impute_formula}} {Either "default" or a path to a yaml file listing the variables to be
                             used as inputs to the imputation model. If a custom file is being used, the top level of the
                             yaml file should be "base" or the name of a variable being imputed with the elements being the
                             names of the variables being used for imputation.}
  \item{\code{standardize}} {A boolean. If TRUE, airpred will perform Z score standardization on
                        specified variables}
  \item{\code{finalday}} {The date of the last day covered by the data set}
  \item{\code{csv_path}} {The path where the assembled data is stored as a csv}
  \item{\code{rds_path}} {The path where the assembled data is stored as an rds file}
  \item{\code{date_var}} {The name of the variable containing date identification}
  \item{\code{site_var}} {the name of the variable containing site identification}
  \item{\code{output_var}} {The name of the variable that the model is trying to predict}
  \item{\code{imputation_models}} {The path where the imputation models should be saved.}
  \item{\code{mid_process_data}} {The path where data should be saved between imputation, normalization
                                  and transformation steps}
  \item{\code{training_models}} {A list of the models to be used in training and used for the
                                 ensemble model.}
  \item{\code{two_stage}} {Should the two stage modeling process be implemented?}
  \item{\code{monitor_list}} {The location of the file containing the coordinates of the monitors}
  \item{\code{training_data}} {The file containing transformed and imputed code to be used for training.
                                Currently must be an RDS file.}
  \item{\code{training_output}} {The directory to be used for storing the output of the training models}
  \item{\code{predict_data}} {The input data for a given round of prediction}
  \item{\code{predict_mid_process}} {The directory that holds all saved files
                                     generated in the prediction process.}
  \item{\code{predict_output}} {The directory that holds the generated predictions}
  \item{\code{pre_generated_weights}} {A boolean determining whether or not the spatial weights
                                      are stored on disk or need to be generated on the fly from
                                      the list of monitors}
   \item{\code{weight_matrix_path}} {The path to where pregenerated weights are stored}
 }
}
