% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{gen_config}
\alias{gen_config}
\title{Generate Config File Skeleton}
\usage{
gen_config(default = TRUE, path = ".", in_list = NULL)
}
\arguments{
\item{default}{A boolean determining whether or not
default values should be loaded into the generated config file}

\item{path}{the directory that the config file should be saved in}

\item{in_list}{An optional list that can be passed in to generate values for a config file.
Currently, the list is not validated, so the config file is not guaranteed to work with
the structure.}
}
\value{
Null, but saves a yml file with the headers needed to run the prediction model saved
}
\description{
Generate Config File Skeleton
}
\details{
The following are the items contained in the config file. All of them must be present in order for
the model to run successfully.

\itemize{
  \item{\code{Monitor}} {The pollution type the data will be trained on}
  \item{\code{Data_Location}} {The directory holding the required data files}
  \item{\code{Data_Save_Location}} {The directory processed data files
                                    should be saved in}
  \item{\code{train}} {A boolean. If TRUE, the model run is a training run. If false,
                       the run is going to be used to create predictions}
  \item{\code{finalday}} {The date of the last day covered by the data set}
  \item{\code{csv_path}} {The path where the assembled data is stored as a csv}
  \item{\code{Imputation_Models}} {The path where the imputation models should be saved.}
  \item{\code{Mid_Process_Data}} {The path where data should be saved between imputation, normalization
                                  and transformation steps}
  \item{\code{Training_Models}} {A list of the models to be used in training and used for the
                                 ensemble model.}
 }
}
