% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_params.R
\name{gen_model_config}
\alias{gen_model_config}
\title{Generate a config file for a given function}
\usage{
gen_model_config(func, path = ".")
}
\arguments{
\item{func}{the function for which a config file is being generated}

\item{path}{the path in which to store the config file}
}
\description{
Generate a config file for a given function
}
\details{
This is used to generate config files for functions not specially implemented
within the airpred. A yaml file with the name "[func]_params.yml" will be
created in your current directory. The file will have its upper level be all of the
tuning arguments for a given function while the lower level will contain the default
values.

NOTE: The function MUST be the first argument passed, otherwise the yaml file will not
be appropriately named, and may end up in an unintended location.
}
