% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenWeights.R
\name{gen_weights}
\alias{gen_weights}
\title{Generate geographic weighting matrix}
\usage{
gen_weights(orig = load_site_list(), query = load_site_list(), k = 4,
  threshold = 1e+05, term = 1)
}
\arguments{
\item{orig}{the data in which the neighbors are being found}

\item{query}{the data for which the neighbors need to be found}

\item{k}{the number of neighbors being searched for}

\item{threshold}{the maximum distance for a neighbor to be given a weight}

\item{term}{the power that the distance is raised to when the weight matrix
is generated.}
}
\description{
Given two sets of points, generate a matrix of weights for the k nearest neighbors
for use in distance weighted averaging of values.
}
