% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{impute_all}
\alias{impute_all}
\title{Impute full dataset}
\usage{
impute_all(info)
}
\arguments{
\item{info}{the dataset being imputed}
}
\value{
data.table with replaced missing values
}
\description{
Impute full dataset
}
