% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{process_data}
\alias{process_data}
\title{Process all data for a given data set}
\usage{
process_data(path = "../predictions/EPANO2")
}
\arguments{
\item{path}{}
}
\description{
Process all data for a given data set
}
