% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{process_data}
\alias{process_data}
\title{Process all data for a given data set}
\usage{
process_data()
}
\value{
none
}
\description{
Process all data for a given data set
}
\details{
saves RDS files into the folder specified in the config file
}
