% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train.R
\name{train_gradboost}
\alias{train_gradboost}
\title{Train Gradient Boost}
\usage{
train_gradboost(info)
}
\arguments{
\item{info}{data frame with model data}
}
\value{
h2o model
}
\description{
Train Gradient Boost
}
