% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train.R
\name{train_gradboost}
\alias{train_gradboost}
\title{Train Gradient Boost}
\usage{
train_gradboost(info, train_ind)
}
\arguments{
\item{info}{data frame with model data}

\item{train_ind}{vector with randomly selected indicies for use as the training set}
}
\value{
h2o model
}
\description{
Train Gradient Boost
}
