% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train.R
\name{train_nn}
\alias{train_nn}
\title{Train Neural Net}
\usage{
train_nn(info, model_name)
}
\arguments{
\item{info}{data frame with model data}

\item{model_name}{name with which to save the model}
}
\value{
h2o model
}
\description{
Train Neural Net
}
